prereq: Use check-lxdialog.sh to determine curses library
Curses detection in adk/config/Makefile is a mess - pkg-config was
designed to avoid exactly that, so make it a required tool on all host
OSs and call check-lxdialog.sh from mconf sources which uses it to
define the required curses LIBS and CFLAGS from prereq.sh.

Since pkg-config output also contains the right arguments for libtinfo
(which depends on the chosen curses variant), no mismatches between the
two may happen anymore.

While being at it, sync check-lxdialog.sh with current linux sources so
it covers installations with widechar support as well.

Signed-off-by: Phil Sutter <phil@nwl.cc>