scripts: Review patch_git.sh and update-patches-git
Fix git-backed package patching and patch updating scripts:

* patch_git.sh was entirely broken (since forever?!) due to a missing
  'cd' call
* update-patches-git formatted the marker commit as patch instead of
  ignoring it

The major improvement of this review is the elimination of the mandatory
final marker commit. Putting it before the series being applied works
well and allows to add a patch to the series without demanding a
git-rebase call to get the marker back on top.

While being at it:

* Call 'git am --abort' only if there was a git repo already
* Call git-commit and git-format-patch with extra options to avoid
  side-effects from a user's .gitconfig
* Explicitly specify output format in git-log to avoid surprises, also
  split output properly with 'read'
* Eliminate fake_hdr variable

Signed-off-by: Phil Sutter <phil@nwl.cc>