package: cfgfs: implement 'diff' command
Sometimes it is unclear how exactly files changed in /etc and the md5sum
is no help there. Implement a diff command for these situations. It is
pretty inefficient since it has to perform a second 'setup' before the
actual 'diff' call. To cover for missing recursion support (and also
exclude internal files) by calling diff for each file separately.

Signed-off-by: Phil Sutter <phil@nwl.cc>