target: linux: config: Set COMPAT_32BIT_TIME on 32bit architectures
Without it, on these architectures 'lvm pvscan' will busy-loop printing:

| io_getevents failed: No such file or directory

The symbol is defined in kernel's arch/Kconfig with 'default !64BIT ||
COMPAT'. Introduce Config.in.compat which has the symbol default to y
unless a known 64bit ADK_TARGET_ARCH is enabled. Drop the symbol from
all target/*/kernel/* for which the default applies.

On 64bit systems, enabling this symbol is probably useful to provide
comaptibility towards 32bit user space. Not sure if this is a possible
choice in OpenADK, but just in case the symbol was not removed from:

- target/aarch64/kernel/raspberry-pi4-64
- target/mips64/kernel/qemu-mips64
- target/riscv64/kernel/qemu-riscv64

Another case is targets explicitly having the symbol disabled, namely
kvx. Given that the two kernel configs in question are full copies of
the kernel-generated one, they are likely not a deliberate choice and
thus removed as well.

Signed-off-by: Phil Sutter <phil@nwl.cc>